refactor: Move V2 endpoints to V1
