+++ cognee/__init__.py
@@ -45,7 +45,7 @@ from cognee.run_migrations import run_startup_migrations
 # ---------------------------------------------------------------------------
 # V2 memory-oriented API
 # ---------------------------------------------------------------------------
-from .api.v2 import remember, RememberResult, recall, improve, forget, serve, disconnect, visualize
+from .api.v1 import remember, RememberResult, recall, improve, forget, serve, disconnect, visualize
 
 # Tracing / Observability
 from cognee.modules.observability.trace_context import (
