+++ cognee/api/v1/serve/device_auth.py
