+++ cognee/tests/unit/modules/memify_tasks/test_sync_graph_to_session.py
@@ -22,8 +22,8 @@ sync_module = sys.modules["cognee.tasks.memify.sync_graph_to_session"]
 # instead of the submodule on Python ≤3.12.
 _mod_sm = importlib.import_module("cognee.infrastructure.session.get_session_manager")
 _mod_cache = importlib.import_module("cognee.infrastructure.databases.cache.get_cache_engine")
-_pkg_improve = importlib.import_module("cognee.api.v2.improve")
-_mod_query_router = importlib.import_module("cognee.api.v2.recall.query_router")
+_pkg_improve = importlib.import_module("cognee.api.v1.improve")
+_mod_query_router = importlib.import_module("cognee.api.v1.recall.query_router")
 
 _PATCH_GET_REL = "cognee.tasks.memify.sync_graph_to_session.get_relational_engine"
 
@@ -111,15 +111,15 @@ def _get_session_manager_module():
 
 
 def _get_remember_module():
-    return importlib.import_module("cognee.api.v2.remember.remember")
+    return importlib.import_module("cognee.api.v1.remember.remember")
 
 
 def _get_improve_module():
-    return importlib.import_module("cognee.api.v2.improve")
+    return importlib.import_module("cognee.api.v1.improve")
 
 
 def _get_query_router_module():
-    return importlib.import_module("cognee.api.v2.recall.query_router")
+    return importlib.import_module("cognee.api.v1.recall.query_router")
 
 
 # ---------------------------------------------------------------------------
@@ -460,7 +460,7 @@ async def test_remember_passes_session_ids_to_improve():
             AsyncMock(return_value=mock_user),
         ),
     ):
-        from cognee.api.v2.remember.remember import remember
+        from cognee.api.v1.remember.remember import remember
 
         await remember(
             "test data",
@@ -494,7 +494,7 @@ async def test_remember_no_session_ids_skips_in_improve():
             AsyncMock(return_value=mock_user),
         ),
     ):
-        from cognee.api.v2.remember.remember import remember
+        from cognee.api.v1.remember.remember import remember
 
         await remember("test data", self_improvement=True)
 
@@ -509,47 +509,47 @@ async def test_remember_no_session_ids_skips_in_improve():
 
 class TestRememberResult:
     def test_repr_basic(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="test")
         assert "completed" in repr(r)
         assert "test" in repr(r)
 
     def test_repr_includes_elapsed(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="test")
         r.elapsed_seconds = 4.2
         assert "elapsed=4.2s" in repr(r)
 
     def test_repr_includes_error(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="errored", dataset_name="test")
         r.error = "something broke"
         assert "something broke" in repr(r)
 
     def test_bool_completed_is_true(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         assert bool(RememberResult(status="completed", dataset_name="x"))
         assert bool(RememberResult(status="session_stored", dataset_name="x"))
 
     def test_bool_running_is_false(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         assert not bool(RememberResult(status="running", dataset_name="x"))
         assert not bool(RememberResult(status="errored", dataset_name="x"))
 
     def test_done_without_task(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         assert RememberResult(status="completed", dataset_name="x").done is True
         assert RememberResult(status="errored", dataset_name="x").done is True
         assert RememberResult(status="running", dataset_name="x").done is False
 
     def test_done_with_task(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="x")
         mock_task = MagicMock()
@@ -562,7 +562,7 @@ class TestRememberResult:
 
     @pytest.mark.asyncio
     async def test_await_completed_returns_self(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x")
         result = await r
@@ -571,7 +571,7 @@ class TestRememberResult:
     @pytest.mark.asyncio
     async def test_await_background_task(self):
         """Awaiting a result with a background task waits for the task."""
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="x")
         completed = False
@@ -590,7 +590,7 @@ class TestRememberResult:
         assert r.status == "completed"
 
     def test_resolve_extracts_pipeline_info(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="test")
 
@@ -609,7 +609,7 @@ class TestRememberResult:
         assert r.elapsed_seconds >= 0
 
     def test_resolve_extracts_item_info(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="test")
 
@@ -639,7 +639,7 @@ class TestRememberResult:
         assert "abc123" in repr(r)
 
     def test_resolve_detects_error(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="test")
 
@@ -650,7 +650,7 @@ class TestRememberResult:
         assert r.status == "errored"
 
     def test_fail_sets_error_and_elapsed(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="running", dataset_name="test")
         r._fail(ValueError("test error"))
@@ -661,7 +661,7 @@ class TestRememberResult:
     @pytest.mark.asyncio
     async def test_remember_returns_remember_result(self):
         """Blocking remember() returns a properly resolved RememberResult."""
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -678,7 +678,7 @@ class TestRememberResult:
                 AsyncMock(return_value=mock_user),
             ),
         ):
-            from cognee.api.v2.remember.remember import remember
+            from cognee.api.v1.remember.remember import remember
 
             result = await remember("test data")
 
@@ -709,7 +709,7 @@ class TestRememberResult:
                 return_value=mock_sm,
             ),
         ):
-            from cognee.api.v2.remember.remember import remember, RememberResult
+            from cognee.api.v1.remember.remember import remember, RememberResult
 
             result = await remember("test data", session_id="s1")
 
@@ -727,35 +727,35 @@ class TestRememberResult:
 
 class TestRememberResultSessions:
     def test_session_id_property_single(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x", session_ids=["s1"])
         assert r.session_id == "s1"
         assert r.session_ids == ["s1"]
 
     def test_session_id_property_multiple(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x", session_ids=["s1", "s2"])
         assert r.session_id is None
         assert r.session_ids == ["s1", "s2"]
 
     def test_session_id_property_none(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x")
         assert r.session_id is None
         assert r.session_ids is None
 
     def test_repr_single_session(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x", session_ids=["s1"])
         assert "session_id='s1'" in repr(r)
         assert "session_ids" not in repr(r)
 
     def test_repr_multiple_sessions(self):
-        from cognee.api.v2.remember.remember import RememberResult
+        from cognee.api.v1.remember.remember import RememberResult
 
         r = RememberResult(status="completed", dataset_name="x", session_ids=["s1", "s2"])
         assert "session_ids=" in repr(r)
@@ -775,7 +775,7 @@ class TestRememberResultSessions:
                 AsyncMock(return_value=mock_user),
             ),
         ):
-            from cognee.api.v2.remember.remember import remember
+            from cognee.api.v1.remember.remember import remember
 
             result = await remember("test data", session_ids=["s1", "s2"], self_improvement=True)
 
@@ -792,7 +792,7 @@ class TestSearchSession:
     @pytest.mark.asyncio
     async def test_word_boundary_matching(self):
         """'graph' should NOT match 'paragraph'."""
-        from cognee.api.v2.recall.recall import _search_session
+        from cognee.api.v1.recall.recall import _search_session
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -821,7 +821,7 @@ class TestSearchSession:
     @pytest.mark.asyncio
     async def test_multiple_word_ranking(self):
         """Entries matching more query words rank higher."""
-        from cognee.api.v2.recall.recall import _search_session
+        from cognee.api.v1.recall.recall import _search_session
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -855,7 +855,7 @@ class TestSearchSession:
     @pytest.mark.asyncio
     async def test_source_tagging(self):
         """Session results should have _source='session'."""
-        from cognee.api.v2.recall.recall import _search_session
+        from cognee.api.v1.recall.recall import _search_session
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -882,7 +882,7 @@ class TestSearchSession:
     @pytest.mark.asyncio
     async def test_empty_session(self):
         """Empty session returns empty list."""
-        from cognee.api.v2.recall.recall import _search_session
+        from cognee.api.v1.recall.recall import _search_session
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -905,7 +905,7 @@ class TestSearchSession:
     @pytest.mark.asyncio
     async def test_short_words_ignored(self):
         """Single-character words like 'a' and 'I' are skipped."""
-        from cognee.api.v2.recall.recall import _search_session
+        from cognee.api.v1.recall.recall import _search_session
 
         mock_user = MagicMock()
         mock_user.id = "u1"
@@ -940,7 +940,7 @@ def _get_recall_module():
     """Import the recall module by full path to avoid __init__.py name collision."""
     import importlib
 
-    return importlib.import_module("cognee.api.v2.recall.recall")
+    return importlib.import_module("cognee.api.v1.recall.recall")
 
 
 class TestRecallSessionMode:
